Added pypi badge and updated build.command

+++ build.command
@@ -1,7 +1,15 @@
 #!/bin/bash
 
+echo 'Building package...'
+if [ ! -d "dist" ]; then
+  mkdir dist
+  echo 'Created directory dist'
+else
+  rm -rf dist/*
+  echo 'Removed previous distribution from dist'
+fi
+
 rm -rf docs/build/*
-rm -rf dist/*
 
 # Evoke Sphinx to create html and pdf documentation
 cd docs
@@ -10,7 +18,7 @@ cd ..
 
 
 # Creating Python packages
-python3.4 setup.py sdist --formats=bztar,gztar,zip
+python3 setup.py sdist --formats=bztar,gztar,zip
 # python3.4 setup.py sdist --formats=zip
 cd dist
 tar xvfj *.bz2
